+++ Editor/VF/Feature/BoundingBoxFixBuilder.cs
@@ -10,7 +10,7 @@ namespace VF.Feature {
 
     public class BoundingBoxFixBuilder : FeatureBuilder<BoundingBoxFix2> {
 
-        [FeatureBuilderAction((int)FeatureOrder.BoundingBoxFix)]
+        [FeatureBuilderAction(FeatureOrder.BoundingBoxFix)]
         public void ApplyOnClone() {
             var skins = avatarObject.GetComponentsInChildren<SkinnedMeshRenderer>(true);
             foreach (var skin in skins) {
